+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changed
 
 - `mpservice.streamer.Streamer.transform`: parameter `concurrency` used to default to 1 (i.e. no concurrency), now defaults to higher numbers (i.e. with concurrency).
 - `mpservice.mpserver.{Sequential, Ensemble}` were renamed to `SequentialServlet` and `EnsembleServlet` respectively.
+- `mpservice.streamer.Streamer.drain`: return count of elements processed, instead of the tuple of element count and exception count.
 
 Added or enhanced
 -----------------